Add test server package and refactor tests to use it
