+++ pkg/server/app/models_test.go
@@ -5,58 +5,8 @@ import (
 	"testing"
 
 	appb "github.com/luizalabs/teresa/pkg/protobuf/app"
+	"github.com/luizalabs/teresa/pkg/server/test"
 )
-
-// Shamelessly adapted from the standard library
-func deepEqual(x, y interface{}) bool {
-	if x == nil || y == nil {
-		return x == y
-	}
-
-	v1 := reflect.ValueOf(x)
-	v2 := reflect.ValueOf(y)
-
-	return deepValueEqual(v1, v2)
-}
-
-func deepValueEqual(v1, v2 reflect.Value) bool {
-	if !v1.IsValid() || !v2.IsValid() {
-		return v1.IsValid() == v2.IsValid()
-	}
-
-	switch v1.Kind() {
-	case reflect.Slice:
-		if v1.IsNil() != v2.IsNil() {
-			return false
-		}
-		if v1.Len() != v2.Len() {
-			return false
-		}
-		if v1.Pointer() == v2.Pointer() {
-			return true
-		}
-		for i := 0; i < v1.Len(); i++ {
-			if !deepValueEqual(v1.Index(i), v2.Index(i)) {
-				return false
-			}
-		}
-		return true
-	case reflect.Ptr:
-		if v1.Pointer() == v2.Pointer() {
-			return true
-		}
-		return deepValueEqual(v1.Elem(), v2.Elem())
-	case reflect.Struct:
-		for i, n := 0, v1.NumField(); i < n; i++ {
-			if !deepValueEqual(v1.Field(i), v2.Field(i)) {
-				return false
-			}
-		}
-		return true
-	default:
-		return v1.Interface() == v2.Interface()
-	}
-}
 
 // We need to skip EnvVars
 func cmpAppWithCreateRequest(app *App, req *appb.CreateRequest) bool {
@@ -74,7 +24,7 @@ func cmpAppWithCreateRequest(app *App, req *appb.CreateRequest) bool {
 		app.Autoscale,
 	}
 
-	return deepEqual(&tmp, req)
+	return test.DeepEqual(&tmp, req)
 }
 
 func newCreateRequest() *appb.CreateRequest {
@@ -149,7 +99,7 @@ func TestNewInfoResponse(t *testing.T) {
 	}
 
 	resp := newInfoResponse(info)
-	if !deepEqual(info, resp) {
+	if !test.DeepEqual(info, resp) {
 		t.Errorf("expected %v, got %v", info, resp)
 	}
 }
@@ -253,7 +203,7 @@ func cmpAutoscaleWithSetAutoscaleRequest(as *Autoscale, req *appb.SetAutoscaleRe
 		as,
 	}
 
-	return deepEqual(&tmp, req)
+	return test.DeepEqual(&tmp, req)
 }
 
 func TestNewAutoscale(t *testing.T) {
