+++ pkg/server/test/reflect.go
@@ -1,0 +1,54 @@
+package test
+
+import "reflect"
+
+// Shamelessly adapted from the standard library
+func DeepEqual(x, y interface{}) bool {
+	if x == nil || y == nil {
+		return x == y
+	}
+
+	v1 := reflect.ValueOf(x)
+	v2 := reflect.ValueOf(y)
+
+	return deepValueEqual(v1, v2)
+}
+
+func deepValueEqual(v1, v2 reflect.Value) bool {
+	if !v1.IsValid() || !v2.IsValid() {
+		return v1.IsValid() == v2.IsValid()
+	}
+
+	switch v1.Kind() {
+	case reflect.Slice:
+		if v1.IsNil() != v2.IsNil() {
+			return false
+		}
+		if v1.Len() != v2.Len() {
+			return false
+		}
+		if v1.Pointer() == v2.Pointer() {
+			return true
+		}
+		for i := 0; i < v1.Len(); i++ {
+			if !deepValueEqual(v1.Index(i), v2.Index(i)) {
+				return false
+			}
+		}
+		return true
+	case reflect.Ptr:
+		if v1.Pointer() == v2.Pointer() {
+			return true
+		}
+		return deepValueEqual(v1.Elem(), v2.Elem())
+	case reflect.Struct:
+		for i, n := 0, v1.NumField(); i < n; i++ {
+			if !deepValueEqual(v1.Field(i), v2.Field(i)) {
+				return false
+			}
+		}
+		return true
+	default:
+		return v1.Interface() == v2.Interface()
+	}
+}
